Update to Support Kernel 6+ and ARM
The new Ubuntu 22.04 VM runs kernel 6.5 which changed some input
parameters for TCP and UDP receive message calls.  Additionally,
some of the print calls I used referenced a iov length that is
no longer part of the struct.  So I needed to  add in kernel
version checks to allow proper compiling.

The ARM architecture was causing issues during build in the json
helper.  So I needed to modify the installer and json helper to
check the architecture and compile with a different version of gcc.