Update Config Script Use
The config script would get the current kernel version to resolve
various locations, but this would cause a lot of files to change
and be tracked by git.

To resolve this, I now write variables to the files that will each
retrieve the kernel version when run.  This will greatly reduce the
number of file changes between different kernels.